Melhorias no backend

+++ projetofinal/backend/controllers/UserController.php
@@ -45,13 +45,15 @@ class UserController extends Controller
      */
     public function actionIndex()
     {
-        $searchModel = new UserSearch();
-        $dataProvider = $searchModel->search($this->request->queryParams);
+        if(Yii::$app->user->can('acessoBackend')) {
+            $searchModel = new UserSearch();
+            $dataProvider = $searchModel->search($this->request->queryParams);
 
-        return $this->render('index', [
-            'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
-        ]);
+            return $this->render('index', [
+                'searchModel' => $searchModel,
+                'dataProvider' => $dataProvider,
+            ]);
+        }
     }
 
     /**
@@ -62,9 +64,11 @@ class UserController extends Controller
      */
     public function actionView($id)
     {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
+        if(Yii::$app->user->can('acessoBackend')) {
+            return $this->render('view', [
+                'model' => $this->findModel($id),
+            ]);
+        }
     }
 
     /**
@@ -74,76 +78,75 @@ class UserController extends Controller
      */
     public function actionCreate()
     {
-        $user = new User();
-        $morada = new Morada();
+        if(Yii::$app->user->can('acessoBackend')) {
+            $user = new User();
+            $morada = new Morada();
 
-        if ($this->request->isPost) {
-            if ($user->load($this->request->post()) && $morada->load($this->request->post()) ) {
-                $this->password = Yii::$app->request->post('password');
+            if ($this->request->isPost) {
+                if ($user->load($this->request->post()) && $morada->load($this->request->post())) {
+                    $this->password = Yii::$app->request->post('password');
+                    /*$checkMorada = Morada::find()->where(['pais' => $morada->pais, 'cidade' => $morada->cidade,
+                        'rua' => $morada->rua, 'codpost' => $morada->codpost])->one();
 
-                $checkMorada = Morada::find()->where(['pais' => $morada->pais, 'cidade' => $morada->cidade,
-                    'rua' => $morada->rua, 'codpost' => $morada->codpost])->one();
-
-                if($checkMorada){
-                    $idMorada = $checkMorada->id;
-                }else {
+                    if ($checkMorada) {
+                        $idMorada = $checkMorada->id;
+                    } else {
+                        $morada->save(false);
+                        $idMorada = $morada->id;
+                    }*/
                     $morada->save(false);
-                    $idMorada = $morada->id;
+                    $user->idMorada = $morada->id;
+                    $user->setPassword($this->password);
+                    $user->generateAuthKey();
+                    $user->save();
+
+                    $auth = \Yii::$app->authManager;
+                    $role = $auth->getRole('Cliente');
+                    try {
+                        $auth->assign($role, $user->id);
+                    } catch (\Exception $e) {
+                    }
+
+                    return $this->redirect(['view', 'id' => $user->id]);
+
+                } else if ($user->load($this->request->post())) {
+                    $this->password = Yii::$app->request->post('password');
+                    $user->setPassword($this->password);
+                    $user->generateAuthKey();
+                    $user->save();
+
+                    $auth = \Yii::$app->authManager;
+                    switch ($user->tipo) {
+                        case 'Admin':
+                            $role = $auth->getRole('Administrador');
+                            break;
+                        case 'Gestor':
+                            $role = $auth->getRole('Gestor');
+                            break;
+                        case 'Funcionario':
+                            $role = $auth->getRole('Funcionario');
+                            break;
+                    }
+                    try {
+                        $auth->assign($role, $user->id);
+                    } catch (\Exception $e) {
+                    }
+
+                    return $this->redirect(['view', 'id' => $user->id]);
+                } else if (!Model::validateMultiple([$user, $morada])) {
+                    Yii::$app->session->setFlash('error', 'Ocorreu um erro ao criar um utilizador.');
+                    return $this->refresh();
                 }
-
-                $user->idMorada = $idMorada;
-                $user->setPassword($this->password);
-                $user->generateAuthKey();
-                $user->save();
-
-                $auth = \Yii::$app->authManager;
-                //$role = $auth->getRole('Cliente');
-                try {
-                    $auth->assign($role, $user->id);
-                } catch (\Exception $e) {
-                }
-                
-                return $this->redirect(['view', 'id' => $user->id]);
-            }else if($user->load($this->request->post())){
-
-                $this->password = Yii::$app->request->post('password');
-                $user->setPassword($this->password);
-                $user->generateAuthKey();
-                $user->save();
-
-                /*
-                $auth = \Yii::$app->authManager;
-                switch ($user->tipo){
-                    case 'Admin':
-                        $role = $auth->getRole('Admin');
-                        break;
-                    case 'Gestor':
-                        $role = $auth->getRole('Gestor');
-                        break;
-                    case 'Funcionario':
-                        $role = $auth->getRole('Funcionario');
-                        break;
-                }
-                try {
-                    $auth->assign($role, $user->id);
-                } catch (\Exception $e) {
-                }*/
-
-                return $this->redirect(['view', 'id' => $user->id]);
+            } else {
+                $user->loadDefaultValues();
+                $morada->loadDefaultValues();
             }
-            else if(!Model::validateMultiple([$user, $morada])){
-                Yii::$app->session->setFlash('error', 'Ocorreu um erro ao criar um utilizador.');
-                return $this->refresh();
-            }
-        } else {
-            $user->loadDefaultValues();
-            $morada->loadDefaultValues();
+
+            return $this->render('create', [
+                'user' => $user,
+                'morada' => $morada,
+            ]);
         }
-
-        return $this->render('create', [
-            'user' => $user,
-            'morada' => $morada,
-        ]);
     }
 
     /**
@@ -155,15 +158,36 @@ class UserController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        if(Yii::$app->user->can('acessoBackend')) {
+            $user = $this->findModel($id);
+            $morada = Morada::findOne($user->idMorada);
 
-        if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+            if ($this->request->isPost) {
+
+                if($morada == null){
+                    if ($user->load($this->request->post()))
+                        $user->save();
+                }else{
+                    if($user->load($this->request->post()) && $morada->load($this->request->post() )){
+                        $morada->save();
+                        $user->save();
+                    }
+                }
+
+                return $this->redirect(['view', 'id' => $user->id]);
+            }
+
+            if($user->tipo == 'Cliente') {
+                return $this->render('formUpdateClientes', [
+                    'user' => $user,
+                    'morada' => $morada,
+                ]);
+            }else{
+                return $this->render('formUpdate', [
+                    'user' => $user,
+                ]);
+            }
         }
-
-        return $this->render('update', [
-            'model' => $model,
-        ]);
     }
 
     /**
@@ -175,9 +199,15 @@ class UserController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
-
-        return $this->redirect(['index']);
+        if(Yii::$app->user->can('acessoBackend')) {
+                $auth = \Yii::$app->authManager;
+                $auth->revokeAll($id);
+                $morada = Morada::findOne($this->findModel($id)->idMorada);
+                $this->findModel($id)->delete();
+                if($morada != null)
+                    $morada->delete();
+                return $this->redirect(['index']);
+        }
     }
 
     /**
@@ -189,10 +219,12 @@ class UserController extends Controller
      */
     protected function findModel($id)
     {
-        if (($model = User::findOne(['id' => $id])) !== null) {
-            return $model;
-        }
+        if(Yii::$app->user->can('acessoBackend')) {
+            if (($model = User::findOne(['id' => $id])) !== null) {
+                return $model;
+            }
 
-        throw new NotFoundHttpException('The requested page does not exist.');
+            throw new NotFoundHttpException('The requested page does not exist.');
+        }
     }
 }
